# backend-laravel/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApiPost - linuxitos",
        "description": "API con login, registro, actualización de información de usuario, perfil, listar posts, agregar, actualizar y eliminar posts. Obtiene token de inicio de sesión y se valida temporalidad.",
        "contact": {
            "email": "contact@linuxitos.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/module": {
            "post": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Obtener lista o un módulo específico",
                "operationId": "5aaf14329038753326235ff3a095fc47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID del módulo",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "module_id",
                        "in": "query",
                        "description": "ID del módulo padre",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de módulos obtenido correctamente"
                    },
                    "401": {
                        "description": "Token inválido o expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/module/delete": {
            "post": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Eliminar un módulo por ID",
                "operationId": "3bd334b88a8b40a345294297f374f3cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID del módulo a eliminar",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Módulo eliminado correctamente"
                    },
                    "404": {
                        "description": "Módulo no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/module/save": {
            "post": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Crear o actualizar un módulo",
                "operationId": "0101a854f64f8dfe5a36e50376bde6e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Dashboard"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Módulo principal"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "fa-dashboard"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "menu"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/dashboard"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#ff6600"
                                    },
                                    "show": {
                                        "type": "string",
                                        "example": "panel"
                                    },
                                    "active": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "module_id": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Módulo creado o actualizado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}